make messages.txt
